+++ SlipeServer.Packets/Definitions/Lua/Rpc/Destroys/DestroyAllPickupsRpcPacket.cs
@@ -1,12 +1,13 @@
 ﻿using SlipeServer.Packets.Builder;
+using SlipeServer.Packets.Definitions.Lua.ElementRpc;
 using SlipeServer.Packets.Enums;
 using System;
 
-namespace SlipeServer.Packets.Definitions.Lua.ElementRpc.Pickups;
+namespace SlipeServer.Packets.Definitions.Lua.Rpc.Destroys;
 
 public class DestroyAllPickupsRpcPacket : Packet
 {
-    public override PacketId PacketId => PacketId.PACKET_ID_LUA_ELEMENT_RPC;
+    public override PacketId PacketId => PacketId.PACKET_ID_LUA;
     public override PacketReliability Reliability => PacketReliability.ReliableSequenced;
     public override PacketPriority Priority => PacketPriority.High;
