+++ SlipeServer.Packets/Definitions/Lua/Rpc/Destroys/DestroyAllVehiclesRpcPacket.cs
@@ -1,12 +1,13 @@
 ﻿using SlipeServer.Packets.Builder;
+using SlipeServer.Packets.Definitions.Lua.ElementRpc;
 using SlipeServer.Packets.Enums;
 using System;
 
-namespace SlipeServer.Packets.Definitions.Lua.ElementRpc.Marker;
+namespace SlipeServer.Packets.Definitions.Lua.Rpc.Destroys;
 
-public class DestroyAllMarkersRpcPacket : Packet
+public class DestroyAllVehiclesRpcPacket : Packet
 {
-    public override PacketId PacketId => PacketId.PACKET_ID_LUA_ELEMENT_RPC;
+    public override PacketId PacketId => PacketId.PACKET_ID_LUA;
     public override PacketReliability Reliability => PacketReliability.ReliableSequenced;
     public override PacketPriority Priority => PacketPriority.High;
 
@@ -18,7 +19,7 @@ public class DestroyAllMarkersRpcPacket : Packet
     public override byte[] Write()
     {
         var builder = new PacketBuilder();
-        builder.Write((byte)ElementRpcFunction.DESTROY_ALL_MARKERS);
+        builder.Write((byte)ElementRpcFunction.DESTROY_ALL_VEHICLES);
         return builder.Build();
     }
 }
