+++ SlipeServer.Server/PacketHandling/Factories/MarkerPacketFactory.cs
@@ -1,4 +1,4 @@
-﻿using SlipeServer.Packets.Definitions.Lua.ElementRpc.Marker;
+﻿using SlipeServer.Packets.Definitions.Lua.Rpc.Destroys;
 
 namespace SlipeServer.Server.PacketHandling.Factories;
