+++ SlipeServer.Server/PacketHandling/Factories/VehiclePacketFactory.cs
@@ -2,6 +2,7 @@
 using SlipeServer.Server.Elements;
 using SlipeServer.Packets.Definitions.Lua.ElementRpc.Vehicle;
 using SlipeServer.Packets.Definitions.Vehicles;
+using SlipeServer.Packets.Definitions.Lua.Rpc.Destroys;
 
 namespace SlipeServer.Server.PacketHandling.Factories;
