+++ SlipeServer.Server/PacketHandling/Factories/WorldObjectPacketFactory.cs
@@ -1,5 +1,6 @@
 ﻿using SlipeServer.Packets.Definitions.Lua.ElementRpc.Element;
 using SlipeServer.Packets.Definitions.Lua.ElementRpc.WorldObject;
+using SlipeServer.Packets.Definitions.Lua.Rpc.Destroys;
 using SlipeServer.Server.Elements;
 
 namespace SlipeServer.Server.PacketHandling.Factories;
